requirements.txt
flake8>=7.0.0,<7.1.0
black>=24.3.0,<24.4.0
jsonschema>=4.21.0,<4.21.1
pystac>=1.9.0,<1.10.0
rasterio>=1.3.9,<1.3.10
shapely>=2.0.3,<2.0.4
azure-storage-blob>=12.19.1,<12.20.0
gdal>=3.8.5,<3.9.0
pydantic-settings>=2.2.1,<2.3.0

